add some api

+++ Project/src/main/java/Zomato/Project/service/MenuItemVariantService.java
@@ -1,6 +1,7 @@
 package Zomato.Project.service;
 
 import Zomato.Project.dto.CombineMenuItemAndMenuItemVariantRequestDTO;
+import Zomato.Project.entity.MenuItem;
 import Zomato.Project.entity.MenuItemVariant;
 import Zomato.Project.entity.Restaurant;
 import Zomato.Project.repository.MenuItemVariantRepository;
@@ -8,6 +9,7 @@ import Zomato.Project.repository.RestaurantRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.List;
 import java.util.Optional;
 
 @Service
@@ -45,6 +47,16 @@ public class MenuItemVariantService {
         if (existingMenuItemVariant.isEmpty()) {
             return "does not exist menu Item variant";
         }
+        MenuItemVariant menuItemVariant = existingMenuItemVariant.get();
+        MenuItem menuItem = menuItemVariant.getMenuItem();
+
+        List<MenuItemVariant> menuItemVariantList = menuItem.getMenuItemVariantList();
+
+        if (menuItemVariantList.size() == 1) {
+            return "cannot delete last variant, at least one variant is required ";
+        }
+
+
         menuItemVariantRepository.deleteById(menuItemVariantId);
         return "Successfully your Menu item Variant is deleted";
     }
